Fix wasm compilation.
