+++ crates/mcp/Cargo.toml
@@ -17,19 +17,23 @@ log.workspace = true
 oauth2.workspace = true
 pin-project-lite = "0.2"
 reqwest = { workspace = true, features = ["json", "stream"] }
-rmcp = { workspace = true, features = ["client", "client-side-sse", "auth", "transport-streamable-http-client-reqwest", "transport-child-process"] }
+rmcp = { workspace = true, features = ["client"] }
 serde.workspace = true
 serde_json.workspace = true
 simple_logger.workspace = true
 sse-stream = "0.2"
 thiserror.workspace = true
-tokio = { workspace = true, features = ["sync", "macros", "rt", "time", "process", "io-util"] }
+tokio.workspace = true
 tracing.workspace = true
 url.workspace = true
 uuid.workspace = true
 warp_core.workspace = true
 warpui.workspace = true
 warpui_extras = { workspace = true, features = ["default", "user_preferences-toml"] }
+
+[target.'cfg(not(target_family = "wasm"))'.dependencies]
+rmcp = { workspace = true, features = ["client", "client-side-sse", "auth", "transport-streamable-http-client-reqwest", "transport-child-process"] }
+tokio = { workspace = true, features = ["sync", "macros", "rt", "time", "process", "io-util"] }
 
 [target.'cfg(windows)'.dependencies]
 windows = { workspace = true, features = ["Win32_System_Threading"] }
